+++ gac_requirement.txt
@@ -1,0 +1,11 @@
+# github actions requirements
+requests
+pandas
+pyqt5
+
+# Linting
+black
+flake8
+flake8-pyproject
+pylint
+
